pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robometrics"
version = "0.1.0"
description = "Scenario-based offline evaluation for robotics logs."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
  { name = "Nitish Sanghi" }
]
keywords = ["robotics", "evaluation", "logs", "research"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Robotics"
]
dependencies = [
  "pandas>=2.2",
  "pyarrow>=14.0",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "ruff>=0.5",
  "black>=24.0"
]

[project.scripts]
robometrics = "robometrics.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88
